unified tests and refactor
